+++ great_international/models/find_a_supplier.py
@@ -11,13 +11,12 @@ import great_international.panels.find_a_supplier as panels
 
 
 class InternationalTradeHomePage(
-    ExclusivePageMixin,
-    ServiceHomepageMixin,
-    BreadcrumbMixin,
-    BaseInternationalPage,
-    panels.InternationalTradeHomePagePanels
+    panels.InternationalTradeHomePagePanels, ExclusivePageMixin, ServiceHomepageMixin,
+    BreadcrumbMixin, BaseInternationalPage,
 ):
     slug_identity = slugs.FIND_A_SUPPLIER_LANDING
+    parent_page_types = ['great_international.InternationalHomePage']
+    subpage_types = ['InternationalTradeIndustryContactPage']
 
     hero_image = models.ForeignKey(
         'wagtailimages.Image',
@@ -83,12 +82,10 @@ class InternationalTradeHomePage(
 
 
 class InternationalTradeIndustryContactPage(
-    ExclusivePageMixin,
-    BreadcrumbMixin,
+    panels.InternationalTradeIndustryContactPagePanels, ExclusivePageMixin, BreadcrumbMixin,
     BaseInternationalPage,
-    panels.InternationalTradeIndustryContactPagePanels,
 ):
-
+    parent_page_types = ['InternationalTradeHomePage']
     view_path = 'industries/contact/'
     slug_identity = slugs.FIND_A_SUPPLIER_INDUSTRY_CONTACT
     # override the slug when generating the url