+++ CHANGELOG.md
@@ -3,16 +3,38 @@
 ## Pre-release
 
 ### Implemented enhancements
-- XOT-932 - Add Country Guide pages to search
-- CI-342 - `CapitalInvestOpportunityPage` gets all projects that have the same related sector as self so they can be randomised
-- CI-344 - `Similar projects` title and `related_page_one/teo/three `CapitalInvestOpportunityPage` are no longer used, only removed from panels for now
+- CMS-1664 (Follow-up) - Update admin panels on international invest homepage
+- CMS-1664 (Follow-up) - Update fields on international invest homepage to match those in invest/models
+- No ticket - Split up code in great domestic to make files smaller
+- No ticket - Update db template after new migrations added
+- CMS-1692 - More small tweaks to Invest HPO pages
+- CI-323 - Added "About the UK" section pages - landing page and "why choose the UK" page
+- CI-365 - Changed the "Contact Us" section on the services page in "About DIT" to have its body be markdown
+- TT-1689 - Move tests in the tests folder
+- CI-368 - Add ebooks promotional section on "why choose the UK" page
+- CI-367 - Update `CapitalInvestRegionPage`
+- CI-366 - Add CTA to UK benefits section of invest home page
+- CI-383 - Populated About DIT landing page
+- CI-325 - Removed prioritised checkbox from capital invest opportunity page as it is now being randomised
+- CI-328 - Removed unused fields for capital invest opportunity page
+- No-ticket - Fix models hierarchy
 
 
 ### Fixed Bugs
+- CI-373 - Fix panels on why choose UK page
 
 
-## [2019.07.18](https://github.com/uktrade/directory-cms/releases/tag/2019.07.18]
-[Full Changelog](https://github.com/uktrade/directory-cms/compare/2019.06.27...2019.07.18
+## [2019.07.24](https://github.com/uktrade/directory-cms/releases/tag/2019.07.24)
+[Full Changelog](https://github.com/uktrade/directory-cms/compare/2019.07.18...2019.07.24)
+
+- XOT-932 - Add Country Guide pages to search
+- CI-342 - CapitalInvestOpportunityPage gets all projects that have the same related sector as self so they can be randomised
+- CI-344 - Similar projects title and related_page_one/teo/three CapitalInvestOpportunityPage` are no longer used, only removed from panels for now
+- CMS-1648 - Added tree-based routing helpers
+
+
+## [2019.07.18](https://github.com/uktrade/directory-cms/releases/tag/2019.07.18)
+[Full Changelog](https://github.com/uktrade/directory-cms/compare/2019.06.27...2019.07.18)
 
 ### Implemented enhancements
 - XOT-922 - Added Marketing Article Page template
@@ -25,7 +47,6 @@
 - CI-324 - Made `BaseInternationalSectorPage` and sector and new sub-sector page inherit from this and opportunity listing page filters by sub sector
 - CI-322 - Added "About DIT" placeholder landing page
 - CI-322 - Added "About DIT" services pages
-- CMS-1648 - Added tree-based routing helpers
 
 ### Fixed Bugs
 - NOTICKET - Capital invest landing page image serializers use aspect ratio for card images
@@ -34,7 +55,7 @@
 
 
 ## [2019.06.27](https://github.com/uktrade/directory-cms/releases/tag/2019.06.27)
-[Full Changelog](https://github.com/uktrade/directory-cms/compare/2019.05.20...2019.06.27
+[Full Changelog](https://github.com/uktrade/directory-cms/compare/2019.05.20...2019.06.27)
 
 ### Implemented enhancements
 
@@ -50,7 +71,7 @@
 
 
 ## [2019.06.20](https://github.com/uktrade/directory-cms/releases/tag/2019.06.20)
-[Full Changelog](https://github.com/uktrade/directory-cms/compare/2019.05.12...2019.06.20
+[Full Changelog](https://github.com/uktrade/directory-cms/compare/2019.05.12...2019.06.20)
 
 **Implemented enhancements:**
 
@@ -74,7 +95,7 @@
 
 
 ## [2019.06.12](https://github.com/uktrade/directory-cms/releases/tag/2019.06.12)
-[Full Changelog](https://github.com/uktrade/directory-cms/compare/2019.05.09...2019.06.12
+[Full Changelog](https://github.com/uktrade/directory-cms/compare/2019.05.09...2019.06.12)
 
 **Implemented enhancements:**